% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexClient.R
\name{loadQuery}
\alias{loadQuery}
\title{Load the result of a query from a database}
\usage{
loadQuery(x, sqltext, ...)
}
\arguments{
\item{x}{a SQLFlexClient object}

\item{sqltext}{a character, the query text}

\item{...}{Additional parameters to dbGetQuery.}
}
\value{
A data.frame
}
\description{
Call the readQuery method of a SQLFlexClient object and retrieve the results.
}
