% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexClient.R
\name{showColumns}
\alias{showColumns}
\title{List all columns of a table}
\usage{
showColumns(x, table_name, schema_name = "public")
}
\arguments{
\item{x}{a SQLFlexClient object}
}
\description{
List all columns of a table
}
