% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexClient.R
\name{showTables}
\alias{showTables}
\title{List all tables accessible through a SQLFlex object}
\usage{
showTables(x)
}
\arguments{
\item{x}{a SQLFlexClient object}
}
\description{
List all tables accessible through a SQLFlex object
}
