% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexClient.R
\docType{class}
\name{SQLFlexClient}
\alias{SQLFlexClient}
\title{SQL database resource client}
\format{
A R6 object of class SQLFlexClient
}
\description{
SQL database resource client

SQL database resource client
}
\details{
Resource client that connects to a SQL database supported by DBI
}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{SQLFlexClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SQLFlexClient$new()}}
\item \href{#method-getConnection}{\code{SQLFlexClient$getConnection()}}
\item \href{#method-readQuery}{\code{SQLFlexClient$readQuery()}}
\item \href{#method-close}{\code{SQLFlexClient$close()}}
\item \href{#method-clone}{\code{SQLFlexClient$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame">}\href{../../resourcer/html/ResourceClient.html#method-asDataFrame}{\code{resourcer::ResourceClient$asDataFrame()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl">}\href{../../resourcer/html/ResourceClient.html#method-asTbl}{\code{resourcer::ResourceClient$asTbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile">}\href{../../resourcer/html/ResourceClient.html#method-downloadFile}{\code{resourcer::ResourceClient$downloadFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec">}\href{../../resourcer/html/ResourceClient.html#method-exec}{\code{resourcer::ResourceClient$exec()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource">}\href{../../resourcer/html/ResourceClient.html#method-getResource}{\code{resourcer::ResourceClient$getResource()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a SQLFlexClient from a resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexClient$new(resource, dbi.connector = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The resource object.}

\item{\code{dbi.connector}}{An optional DBIResourceConnector object. If not provided, it will be looked up in the DBIResourceConnector registry.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The SQLResourceClient object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
Get or create the DBI connection object that will access the resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexClient$getConnection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The DBI connection object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readQuery"></a>}}
\if{latex}{\out{\hypertarget{method-readQuery}{}}}
\subsection{Method \code{readQuery()}}{
Execute a query in the database and retrieve the results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexClient$readQuery(sqltext, params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sqltext}}{a character, the query text}

\item{\code{params}}{Additional parameters to dbGetQuery.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Silently close the DBI connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
