% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexResolver.R
\docType{class}
\name{SQLFlexResolver}
\alias{SQLFlexResolver}
\title{The resource is SQL database.}
\format{
A R6 object of class SQLResourceResolver
}
\description{
The resource is SQL database.

The resource is SQL database.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{SQLFlexResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-isFor}{\code{SQLFlexResolver$isFor()}}
\item \href{#method-newClient}{\code{SQLFlexResolver$newClient()}}
\item \href{#method-clone}{\code{SQLFlexResolver$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize">}\href{../../resourcer/html/ResourceResolver.html#method-initialize}{\code{resourcer::ResourceResolver$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check that the provided resource has a registered DBIResourceConnector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexResolver$isFor(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The resource object to evaluate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-newClient}{}}}
\subsection{Method \code{newClient()}}{
Creates a SQLFlexClient instance from provided resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexResolver$newClient(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A valid resource object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A SQLFlexClient object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLFlexResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
