% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexClient.R
\name{viewSize}
\alias{viewSize}
\title{Estimate the memory size of one or more views}
\usage{
viewSize(db, views)
}
\arguments{
\item{db}{a SQLFlexClient object}

\item{views}{a character vector, names of the views (or sql queries)}
}
\description{
Estimate the memory size of one or more views
}
