% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexClient.R
\name{loadQuery}
\alias{loadQuery}
\title{wrapper for qLoad, takes care of some parsing}
\usage{
loadQuery(x, table_name, cols, where_clause, params = NULL)
}
\description{
wrapper for qLoad, takes care of some parsing
}
