% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLFlexClient.R
\name{showColumns}
\alias{showColumns}
\title{List all columns of a table}
\usage{
showColumns(x, table_name)
}
\arguments{
\item{x}{a SQLFlexClient object}

\item{table_name}{name of the table (maybe qualified with schema name)}
}
\description{
List all columns of a table
}
